requirements.txt
# Optional: Uncomment if you want to use Pydantic for validation
# pydantic>=2.0.0
pypdf2==3.0.1
pillow==10.1.0
pytesseract==0.3.10
groq==0.4.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
redis
PyPDF2==3.0.1
python-docx==1.1.0

